remove renaming changes from benchmark package til after that PR is up to avoid conflicts
