+++ src/guidellm/__main__.py
@@ -6,7 +6,7 @@ from typing import get_args
 import click
 from pydantic import ValidationError
 
-from guidellm.backends import BackendType
+from guidellm.backend import BackendType
 from guidellm.benchmark import (
     ProfileType,
     reimport_benchmarks_report,
